recipebook/requirements.txt
asgiref==3.3.1
astroid==2.5
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.3
mccabe==0.6.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
